Floating point probabilities

+++ otherblocks/src/com/sargant/bukkit/otherblocks/OtherBlocksContainer.java
@@ -9,6 +9,6 @@ public class OtherBlocksContainer
 	public Material tool;
 	public Integer quantity;
 	public Integer damage;
-	public Integer chance;
+	public Double chance;
 	public short color;
 }
